requirements.txt
fastapi
uvicorn[standard]
google-cloud-aiplatform
google-auth
pydantic
